+++ client/components/OrderSuccess.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const OrderSuccess = () => {
+  return (
+    <div>
+      <h2>Your order has been successfully processed. Thanks for coming by!</h2>
+    </div>
+  );
+};
+
+export default OrderSuccess;
