+++ client/store/index.js
@@ -5,13 +5,11 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import user from './user';
 import productReducer from './product';
 import cartReducer from './cart';
-import paymentReducer from './payment';
 
 const reducer = combineReducers({
   user,
   products: productReducer,
   cart: cartReducer,
-  payment: paymentReducer,
 });
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
@@ -22,4 +20,3 @@ export default store;
 export * from './user';
 export * from './product';
 export * from './cart';
-export * from './payment';
